+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## v0.20.0
 
 - TTL serialization now treats IRIs correctly, i.e., they are not enclosed in `<...>`
-- Allow using literal fields (`Literal`)
+- Allow using `LangString` to set language-tagged strings as values of properties
 
 ## v0.19.4